Split test_loop.rs into submodules.
