+++ chain/client/src/test_utils/mod.rs
@@ -2,8 +2,6 @@ pub mod block_stats;
 pub mod client;
 pub mod peer_manager_mock;
 pub mod setup;
-pub mod sync_actor_test_utils;
-pub mod sync_jobs_test_utils;
 mod synchronous_state_witness_adapter;
 pub mod test_env;
 pub mod test_env_builder;
