+++ integration-tests/src/tests/client/features/multinode_test_loop_example.rs
@@ -1,7 +1,7 @@
 use derive_enum_from_into::{EnumFrom, EnumTryInto};
 use near_async::futures::FutureSpawner;
 use near_async::messaging::{
-    noop, CanSend, IntoMultiSender, IntoSender, MessageWithCallback, SendAsync,
+    noop, IntoMultiSender, IntoSender, MessageWithCallback, SendAsync,
 };
 use near_async::test_loop::adhoc::{handle_adhoc_events, AdhocEvent, AdhocEventSender};
 use near_async::test_loop::event_handler::ignore_events;
@@ -42,22 +42,16 @@ use near_client::sync_jobs_actions::{
     ClientSenderForSyncJobsMessage, SyncJobsActions, SyncJobsSenderForSyncJobsMessage,
 };
 use near_client::test_utils::test_loop::{
-    forward_client_messages_from_client_to_client_actions,
-    forward_client_messages_from_network_to_client_actions,
-    forward_client_messages_from_shards_manager, forward_client_messages_from_sync_adapter,
-    forward_client_messages_from_sync_jobs_to_client_actions,
-    forward_messages_from_client_to_state_witness_actor,
-    forward_messages_from_network_to_state_witness_actor,
     print_basic_client_info_before_each_event,
 };
-use near_client::test_utils::sync_actor_test_utils::{
+use near_client::test_utils::test_loop::sync_actor::{
     forward_sync_actor_messages_from_client, forward_sync_actor_messages_from_network,
     test_loop_sync_actor_maker, TestSyncActors,
 };
 use near_client::{
     Client, StateWitnessActions, StateWitnessSenderForClientMessage, SyncAdapter, SyncMessage,
 };
-use near_client::test_utils::sync_jobs_test_utils::{
+use near_client::test_utils::test_loop::sync_jobs_actions::{
     forward_sync_jobs_messages_from_client_to_sync_jobs_actions,
     forward_sync_jobs_messages_from_sync_jobs_to_sync_jobs_actions,
 };
@@ -65,12 +59,11 @@ use near_client::test_utils::test_loop::{route_network_messages_to_client, Clien
 use near_epoch_manager::shard_tracker::{ShardTracker, TrackedConfig};
 use near_epoch_manager::EpochManager;
 use near_network::client::{
-    BlockApproval, BlockResponse, ChunkEndorsementMessage, ChunkStateWitnessMessage,
     ClientSenderForNetwork, ClientSenderForNetworkMessage, ProcessTxRequest,
 };
 use near_network::shards_manager::ShardsManagerRequestFromNetwork;
 use near_network::state_witness::{
-    ChunkStateWitnessAckMessage, StateWitnessSenderForNetwork, StateWitnessSenderForNetworkMessage,
+    StateWitnessSenderForNetworkMessage,
 };
 use near_network::state_sync::StateSyncResponse;
 use near_network::types::{PeerManagerMessageRequest, PeerManagerMessageResponse, SetChainInfo};
@@ -93,6 +86,8 @@ use nearcore::state_sync::StateSyncDumper;
 use nearcore::NightshadeRuntime;
 use std::collections::HashMap;
 use std::sync::{Arc, Mutex, RwLock};
+use near_client::test_utils::test_loop::client_actions::{forward_client_messages_from_client_to_client_actions, forward_client_messages_from_network_to_client_actions, forward_client_messages_from_shards_manager, forward_client_messages_from_sync_adapter, forward_client_messages_from_sync_jobs_to_client_actions};
+use near_client::test_utils::test_loop::state_witness_actions::{forward_messages_from_client_to_state_witness_actor, forward_messages_from_network_to_state_witness_actor};
 
 #[derive(derive_more::AsMut, derive_more::AsRef)]
 struct TestData {
