+++ integration-tests/src/tests/client/features/simple_test_loop_example.rs
@@ -20,11 +20,6 @@ use near_client::client_actions::{
 use near_client::sync_jobs_actions::{
     ClientSenderForSyncJobsMessage, SyncJobsActions, SyncJobsSenderForSyncJobsMessage,
 };
-use near_client::test_utils::test_loop::{
-    forward_client_messages_from_client_to_client_actions,
-    forward_client_messages_from_shards_manager,
-    forward_client_messages_from_sync_jobs_to_client_actions,
-};
 use near_client::test_utils::{MAX_BLOCK_PROD_TIME, MIN_BLOCK_PROD_TIME};
 use near_client::{Client, SyncAdapter, SyncMessage};
 use near_epoch_manager::shard_tracker::{ShardTracker, TrackedConfig};
@@ -44,7 +39,8 @@ use near_store::test_utils::create_test_store;
 use nearcore::NightshadeRuntime;
 use std::path::Path;
 use std::sync::{Arc, RwLock};
-use near_client::test_utils::sync_jobs_test_utils::forward_sync_jobs_messages_from_client_to_sync_jobs_actions;
+use near_client::test_utils::test_loop::client_actions::{forward_client_messages_from_client_to_client_actions, forward_client_messages_from_shards_manager, forward_client_messages_from_sync_jobs_to_client_actions};
+use near_client::test_utils::test_loop::sync_jobs_actions::forward_sync_jobs_messages_from_client_to_sync_jobs_actions;
 
 #[derive(derive_more::AsMut, derive_more::AsRef)]
 struct TestData {
